+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/TestRunnerActivity.java
@@ -34,6 +34,7 @@ import android.view.Surface;
 
 import java.util.ArrayDeque;
 import java.util.HashMap;
+import java.util.List;
 
 public class TestRunnerActivity extends Activity {
     private static final String LOGTAG = "TestRunnerActivity";